change overallStatus type in newFeedbackDto to float.

+++ Models/Dtos/NewPatientFeedbackDto.cs
@@ -5,7 +5,7 @@
              public Guid feedbackId { get; set; }
              public Guid patientId { get; set; }
              public Guid appointmentId { get; set; }
-             public int overallStatus { get; set; }
+             public float overallStatus { get; set; }
              public string newSymptoms { get; set; }
              public string comments { get; set; }
              public DateTime createdOn { get; set; }
